Refactor Node and add descriptive comments
Types is no longer a field, and but the type list still exists in the
constructor. This was done b/c the constructor is the only scope it is used
in.

Also, since a node is assumed to empty by default, the fields were reset
appropriately, and the 'if type == -1' check was removed from the constructor.